Move GraphiQL license attribution into README.
Previously it was appended to `LICENSE.txt` but GitHub's automatic
license detection isn't able to detect that it's an MIT license.

The `licensee` gem, which GitHub uses for license detection, shows the problem:

```
$ licensee detect https://github.com/block/elasticgraph
License:        NOASSERTION
Matched files:  LICENSE.txt
LICENSE.txt:
  Content hash:  73812563c638c8691288d777f43f7d738fef02b1
  License:       NOASSERTION
  Closest non-matching licenses:
    MIT similarity:      75.61%
    NCSA similarity:     66.18%
    BSL-1.0 similarity:  64.38%
$ licensee diff https://github.com/block/elasticgraph --license=MIT
Comparing to MIT License:
Input Length:      1213
License length:    1020
Similarity:      75.61%
diff --git a/LICENSE b/LICENSE
index 49b0d1f..0cf6233 100644
--- a/LICENSE
+++ b/LICENSE
@@ -11,4 +11,7 @@ merchantability, fitness for a particular purpose and noninfringement. in no
event shall the authors or copyright holders be liable for any claim, damages or
other liability, whether in an action of contract, tort or otherwise, arising
from, out of or in connection with the software or the use or other dealings in
the software. {+part of the distributed code+}
{+(elasticgraph-rack/lib/elastic_graph/rack/graphiql/index.html) comes from the+}
{+graphiql project, licensed under the mit license. copyright (c) graphql+}
{+contributors.+}
```

GitHub[^1] recommends putting this extra bit of license info in the README:

> To have your license detected, simplify your LICENSE file and note the complexity somewhere else, such as your repository's README file.

[^1]: https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/licensing-a-repository#detecting-a-license