Update the FullTextSearch docs to provide an example of their optionality (#25)
* Update the FullTextSearch docs to provide an example of their optionality

In order to ignore a `matchesQuery` or `matchesPhrase` filter predicate, you can supply `null` to the `MatchesQueryFilterInput` or `MatchesQueryFilterInput` types, respectively. However, this is not entirely clear from the docs, as they describe supplying `null`, but not to which field/value. This documentation update aims to make that more clear, and provides an example to highlight the ability to use a null-able variable with these predicates.

## Tested
Ran the site locally. Here are the screenshots of updates:

* update PR with suggestions from comments, including:
* Remove the language around "ignoring" filters
* Revert the changes to the high-level fulltext.md file
* Alter the example query to use 2 spaces instead of 4-space tabs