Reorder `filter_node_interpreter` to check `:empty` case first.
We want a filter like `gt: nil` to be categorized as `:empty` rather
than `:operator`, as this allows us to avoid a bunch of empty filter
checks litered throughout `FilterInterpreter`. In addition, we need this
change before we can refactor `FilterValueSetExtractor` to leverage
`FilterNodeInterpreter`--empty filter handling needs to take precedence
there.