Move `UnboundedWithExclusions` into `FilterValueSetExtractor`.
The case of a set which is unbounded but excludes some values is
possible for more than just `RoutingPicker` (where we previously
had it defined). It's possible for any usage of `FilterValueSetExtractor`!

However, we haven't run into this in `IndexExpressionBuilder` because
`FilterValueSetExtractor` has a subtle bug that accidentally works around
this issue. In a later PR, I'm going to be reworking `FilterValueSetExtractor`
to move away from the bug (by leveraging `FilterNodeInterpreter`) and
this bug must be fixed first.